square in the center with mid centre pattern
Write ALP that will display the pattern shown in Figure above in the middle of the screen. The pattern
should not blink. You will have to use text mode for this display. The pattern is a red rectangle (11
rows x 20 Columns) with a bright cyan set of crosses in the centre of the rectangle. The rest of the
screen remains black. ‘?’ should be used as blocking function.