pyramid at the center of the screen
Write ALP that will display the pattern shown in Figure above in the middle of the screen. The pattern
should not blink. You will have to use text mode for this display. The pattern is a green pyramid (10
rows x 19 Columns at the base of the pyramid). The rest of the screen remains black. ‘?’ should be
used as blocking function.