square and star pattern at the centre of screen
Write ALP that will display the pattern shown in Figure above in the middle of the screen. The
pattern should blink. You will have to use text mode for this display. The pattern is blue box (10 rows
x 10columns) filled with white stars. The rest of the screen remains black. A Question mark (‘?’)
should be used to exit from the display